Backend/requirements.txt
fastapi
openai
argparse
uvicorn
google-cloud-secret-manager
faiss-cpu==1.7.4
sentence-transformers
numpy
pickle5